Reforcar integridade de empresas

+++ backend/app/models/customer.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Float, ForeignKey, DateTime, Index, Date, Boolean
+from sqlalchemy import Column, Integer, String, Numeric, ForeignKey, DateTime, Index, Date, Boolean
 from sqlalchemy.orm import relationship
 from datetime import datetime, date
 from app.models.base import Base, TimestampMixin
@@ -12,14 +12,14 @@ class Customer(Base, TimestampMixin):
     telefone = Column(String(20), nullable=True, index=True)  # Index para busca rápida
     email = Column(String(255), nullable=True)
     data_nascimento = Column(Date, nullable=True)  # Para aniversariantes
-    pontos = Column(Float, default=0.0, nullable=False)  # Saldo atual
+    pontos = Column(Numeric(12, 2), default=0, nullable=False)  # Saldo atual
     ativo = Column(Boolean, default=True, nullable=False)  # Flag de atividade
     company_id = Column(Integer, ForeignKey("companies.id"), nullable=False, index=True)
     
     # Rastreamento de progresso para premiação
-    valor_gasto_atual = Column(Float, default=0.0, nullable=False)  # Para promoção por valor
+    valor_gasto_atual = Column(Numeric(12, 2), default=0, nullable=False)  # Para promocao por valor
     quantidade_produtos_comprados = Column(Integer, default=0, nullable=False)  # Para promoção por quantidade
-    meta_premiacao_valor = Column(Float, nullable=True)  # Meta em Reais
+    meta_premiacao_valor = Column(Numeric(12, 2), nullable=True)  # Meta em Reais
     meta_premiacao_quantidade = Column(Integer, nullable=True)  # Meta em quantidade de produtos
 
     # Índices para queries rápidas
